+++ docs/source/changes.rst
@@ -8,7 +8,8 @@ Release 0.043 - (under development)
 * Improve support for latest NVIDIA GPUs.
 * Add more wrappers for CUBLAS 5 functions (enh. by Teodor Moldovan)
 * Add support for CULA Dense Free R17 (enh. by Alex Rubinsteyn)
-
+* Memoize elementwise kernel used by ifft scaling (#37).
+  
 Release 0.042 - (March 10, 2013)
 --------------------------------
 * Add complex exponential integral.